pyproject.toml
[tool.poetry]
name = "auto-bi-assistant"
version = "0.1.0"
description = "Agente AI per estrazione e classificazione spese bancarie"
authors = ["Il tuo nome <la.tua.email@example.com>"]
readme = "README.md"
packages = [{include = "auto_bi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.118.0"
google-auth-oauthlib = "^1.2.0"
pydantic = "^2.6.0"
instructor = "^1.0.0"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
google-auth-httplib2 = "^0.3.1"
beautifulsoup4 = "^4.14.3"
plotly = "^6.7.0"
bs4 = "^0.0.2"
streamlit = "^1.30.0"
ddgs = "^9.13.0"
pandas = "^2.0.0"
openpyxl = "^3.1.2"
watchdog = "^6.0.0"
pdfplumber = "^0.11.4"
fastapi = "^0.136.0"
uvicorn = "^0.46.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.49"}
asyncpg = "^0.31.0"
alembic = "^1.18.4"
celery = "^5.6.3"
redis = "^7.4.0"
python-multipart = "^0.0.26"
greenlet = "^3.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.2.2"

[tool.poetry.scripts]
auto-bi = "auto_bi.main:main_cli"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E722"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"